Move RouteStopsOverlayRow to separate component
